MQTT works in it's entirety now. Have to fix an Expetion 0 issue.

+++ include/CurrentReader.hpp
@@ -13,9 +13,9 @@ class CurrentReader {
 
     private:
         ESP8266Timer timer;
-        static inline const uint16_t lastADCValuesLength = 100;
-        static volatile uint16_t lastADCValues[]; // Stores the last X ADC values. 
-        static volatile uint16_t lastADCValuesIndex ;
+        static inline const int lastADCValuesLength = 100;
+        static volatile int lastADCValues[]; // Stores the last X ADC values. 
+        static volatile int lastADCValuesIndex ;
         static inline const uint16 thresholdForMachineOn = 200; // If the average of the last X ADC values is above this the machine is considered on.
         static void IRAM_ATTR readADC(); // Having IRAM_ATTR in the header maybe breaks things, idk
